requirements.txt
ultralytics
supervision


# Our own extras
natsort
scikit-learn
optuna
fiftyone
# nvidia-dlprof
colorama
jsonpickle

# Base ----------------------------------------
matplotlib>=3.2.2
numpy==1.23.3
opencv-python>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
#torch>=1.7.0,!=1.12.0
#torchvision>=0.8.1,!=0.13.0
tqdm>=4.41.0
protobuf<4.21.3

loguru
easydict

# Logging -------------------------------------
tensorboard>=2.4.1
wandb
torch_tb_profiler

# Plotting ------------------------------------
pandas>=1.1.4
seaborn>=0.11.0

# Export --------------------------------------
coreml
#onnxsim
#nvidia-pyindex
#onnx_graphsurgeon
colored
coremltools>=4.1                # CoreML export
onnx>=1.9.0                     # ONNX export
# onnx-simplifier>=0.3.6          # ONNX simplifier
scikit-learn                    # CoreML quantization
# tensorflow>=2.4.1             # TFLite export
# tensorflowjs>=3.9.0           # TF.js export
# openvino-dev                  # OpenVINO export

# Extras --------------------------------------
ipython                         # interactive notebook
psutil                          # system utilization
thop                            # FLOPs computation
albumentations>=1.0.3           # Perhaps something for data augmentation
pycocotools>=2.0                # For computing COCO mAP